move siteConfig to another file
